@@ -1,7 +1,7 @@
 {
     "name": "symfony/http-foundation",
     "type": "library",
-    "description": "Symfony HttpFoundation Component",
+    "description": "Defines an object-oriented layer for the HTTP specification",
     "keywords": [],
     "homepage": "https://symfony.com",
     "license": "MIT",
@@ -16,13 +16,22 @@
         }
     ],
     "require": {
-        "php": ">=5.3.9",
-        "symfony/polyfill-php54": "~1.0",
-        "symfony/polyfill-php55": "~1.0",
-        "symfony/polyfill-mbstring": "~1.1"
+        "php": ">=7.2.5",
+        "symfony/deprecation-contracts": "^2.1|^3",
+        "symfony/polyfill-mbstring": "~1.1",
+        "symfony/polyfill-php80": "^1.16"
     },
     "require-dev": {
-        "symfony/expression-language": "~2.4|~3.0.0"
+        "predis/predis": "^1.0|^2.0",
+        "symfony/cache": "^4.4|^5.0|^6.0",
+        "symfony/dependency-injection": "^5.4|^6.0",
+        "symfony/http-kernel": "^5.4.12|^6.0.12|^6.1.4",
+        "symfony/mime": "^4.4|^5.0|^6.0",
+        "symfony/expression-language": "^4.4|^5.0|^6.0",
+        "symfony/rate-limiter": "^5.2|^6.0"
+    },
+    "suggest" : {
+        "symfony/mime": "To use the file extension guesser"
     },
     "autoload": {
         "psr-4": { "Symfony\\Component\\HttpFoundation\\": "" },
@@ -30,10 +39,5 @@
             "/Tests/"
         ]
     },
-    "minimum-stability": "dev",
-    "extra": {
-        "branch-alias": {
-            "dev-master": "2.8-dev"
-        }
-    }
+    "minimum-stability": "dev"
 }
