@@ -46,7 +46,7 @@
         "php": ">=7.4",
         "phpmailer/phpmailer": "^6",
         "predis/predis": "^1.1",
-        "symfony/http-foundation": "^2.8.52",
+        "symfony/http-foundation": "^5|^6",
         "symfony/yaml": "^4|^5",
         "amzn/login-and-pay-with-amazon-sdk-php": "^2.5",
         "psr/cache": "^1.0",
