@@ -21,20 +21,18 @@ use Symfony\Component\HttpFoundation\Session\SessionBagInterface;
 interface FlashBagInterface extends SessionBagInterface
 {
     /**
-     * Adds a flash message for type.
+     * Adds a flash message for the given type.
      *
-     * @param string $type
-     * @param mixed  $message
+     * @param mixed $message
      */
-    public function add($type, $message);
+    public function add(string $type, $message);
 
     /**
-     * Registers a message for a given type.
+     * Registers one or more messages for a given type.
      *
-     * @param string       $type
-     * @param string|array $message
+     * @param string|array $messages
      */
-    public function set($type, $message);
+    public function set(string $type, $messages);
 
     /**
      * Gets flash messages for a given type.
@@ -44,7 +42,7 @@ interface FlashBagInterface extends SessionBagInterface
      *
      * @return array
      */
-    public function peek($type, array $default = array());
+    public function peek(string $type, array $default = []);
 
     /**
      * Gets all flash messages.
@@ -56,12 +54,11 @@ interface FlashBagInterface extends SessionBagInterface
     /**
      * Gets and clears flash from the stack.
      *
-     * @param string $type
-     * @param array  $default Default value if $type does not exist
+     * @param array $default Default value if $type does not exist
      *
      * @return array
      */
-    public function get($type, array $default = array());
+    public function get(string $type, array $default = []);
 
     /**
      * Gets and clears flashes from the stack.
@@ -78,11 +75,9 @@ interface FlashBagInterface extends SessionBagInterface
     /**
      * Has flash messages for a given type?
      *
-     * @param string $type
-     *
      * @return bool
      */
-    public function has($type);
+    public function has(string $type);
 
     /**
      * Returns a list of all defined types.
