@@ -3,7 +3,7 @@
         'name' => 'mediawiki/core',
         'pretty_version' => 'dev-master',
         'version' => 'dev-master',
-        'reference' => '6527e320f3a2d68d769ea68a0fa78c4b8619b3ea',
+        'reference' => '06208e4e22f8b305ced42b18b2bc6f6cf2b82702',
         'type' => 'mediawiki-core',
         'install_path' => __DIR__ . '/../../',
         'aliases' => array(),
@@ -202,7 +202,7 @@
         'mediawiki/core' => array(
             'pretty_version' => 'dev-master',
             'version' => 'dev-master',
-            'reference' => '6527e320f3a2d68d769ea68a0fa78c4b8619b3ea',
+            'reference' => '06208e4e22f8b305ced42b18b2bc6f6cf2b82702',
             'type' => 'mediawiki-core',
             'install_path' => __DIR__ . '/../../',
             'aliases' => array(),
@@ -467,9 +467,9 @@
             'dev_requirement' => false,
         ),
         'symfony/http-foundation' => array(
-            'pretty_version' => 'v2.8.52',
-            'version' => '2.8.52.0',
-            'reference' => '3929d9fe8148d17819ad0178c748b8d339420709',
+            'pretty_version' => 'v5.4.46',
+            'version' => '5.4.46.0',
+            'reference' => '168b77c71e6f02d8fc479db78beaf742a37d3cab',
             'type' => 'library',
             'install_path' => __DIR__ . '/../symfony/http-foundation',
             'aliases' => array(),
@@ -486,24 +486,6 @@
             'replaced' => array(
                 0 => '1.99',
             ),
-        ),
-        'symfony/polyfill-php54' => array(
-            'pretty_version' => 'v1.20.0',
-            'version' => '1.20.0.0',
-            'reference' => '37285b1d5d13f37c8bee546d8d2ad0353460c4c7',
-            'type' => 'metapackage',
-            'install_path' => null,
-            'aliases' => array(),
-            'dev_requirement' => false,
-        ),
-        'symfony/polyfill-php55' => array(
-            'pretty_version' => 'v1.20.0',
-            'version' => '1.20.0.0',
-            'reference' => 'c17452124a883900e1d73961f9075a638399c1a0',
-            'type' => 'metapackage',
-            'install_path' => null,
-            'aliases' => array(),
-            'dev_requirement' => false,
         ),
         'symfony/polyfill-php80' => array(
             'pretty_version' => 'v1.28.0',
@@ -794,9 +776,9 @@
             'dev_requirement' => false,
         ),
         'wikimedia/smash-pig' => array(
-            'pretty_version' => 'v0.8.22',
-            'version' => '0.8.22.0',
-            'reference' => 'c9872a146164e0531e6f579f6a4966a80f3d47ad',
+            'pretty_version' => 'v0.8.23',
+            'version' => '0.8.23.0',
+            'reference' => '9f64602ef9b84fcd43eff1abbfbc13e7c21dec4d',
             'type' => 'library',
             'install_path' => __DIR__ . '/../wikimedia/smash-pig',
             'aliases' => array(),
