@@ -35,13 +35,13 @@ class ExpressionRequestMatcher extends RequestMatcher
             throw new \LogicException('Unable to match the request as the expression language is not available.');
         }
 
-        return $this->language->evaluate($this->expression, array(
+        return $this->language->evaluate($this->expression, [
             'request' => $request,
             'method' => $request->getMethod(),
             'path' => rawurldecode($request->getPathInfo()),
             'host' => $request->getHost(),
             'ip' => $request->getClientIp(),
             'attributes' => $request->attributes->all(),
-        )) && parent::matches($request);
+        ]) && parent::matches($request);
     }
 }
