@@ -18,10 +18,7 @@ namespace Symfony\Component\HttpFoundation\File\Exception;
  */
 class AccessDeniedException extends FileException
 {
-    /**
-     * @param string $path The path to the accessed file
-     */
-    public function __construct($path)
+    public function __construct(string $path)
     {
         parent::__construct(sprintf('The file %s could not be accessed', $path));
     }
