@@ -18,10 +18,7 @@ namespace Symfony\Component\HttpFoundation\File\Exception;
  */
 class FileNotFoundException extends FileException
 {
-    /**
-     * @param string $path The path to the file that was not found
-     */
-    public function __construct($path)
+    public function __construct(string $path)
     {
         parent::__construct(sprintf('The file "%s" does not exist', $path));
     }
