@@ -13,8 +13,8 @@ namespace Symfony\Component\HttpFoundation\File\Exception;
 
 class UnexpectedTypeException extends FileException
 {
-    public function __construct($value, $expectedType)
+    public function __construct($value, string $expectedType)
     {
-        parent::__construct(sprintf('Expected argument of type %s, %s given', $expectedType, \is_object($value) ? \get_class($value) : \gettype($value)));
+        parent::__construct(sprintf('Expected argument of type %s, %s given', $expectedType, get_debug_type($value)));
     }
 }
