@@ -21,7 +21,7 @@ interface RequestMatcherInterface
     /**
      * Decides whether the rule(s) implemented by the strategy matches the supplied request.
      *
-     * @return bool true if the request matches, false otherwise
+     * @return bool
      */
     public function matches(Request $request);
 }
