@@ -23,50 +23,39 @@ interface AttributeBagInterface extends SessionBagInterface
     /**
      * Checks if an attribute is defined.
      *
-     * @param string $name The attribute name
-     *
-     * @return bool true if the attribute is defined, false otherwise
+     * @return bool
      */
-    public function has($name);
+    public function has(string $name);
 
     /**
      * Returns an attribute.
      *
-     * @param string $name    The attribute name
-     * @param mixed  $default The default value if not found
+     * @param mixed $default The default value if not found
      *
      * @return mixed
      */
-    public function get($name, $default = null);
+    public function get(string $name, $default = null);
 
     /**
      * Sets an attribute.
      *
-     * @param string $name
-     * @param mixed  $value
+     * @param mixed $value
      */
-    public function set($name, $value);
+    public function set(string $name, $value);
 
     /**
      * Returns attributes.
      *
-     * @return array Attributes
+     * @return array<string, mixed>
      */
     public function all();
 
-    /**
-     * Sets attributes.
-     *
-     * @param array $attributes Attributes
-     */
     public function replace(array $attributes);
 
     /**
      * Removes an attribute.
      *
-     * @param string $name
-     *
      * @return mixed The removed value or null when it does not exist
      */
-    public function remove($name);
+    public function remove(string $name);
 }
