@@ -24,7 +24,7 @@ interface SessionStorageInterface
     /**
      * Starts the session.
      *
-     * @return bool True if started
+     * @return bool
      *
      * @throws \RuntimeException if something goes wrong starting the session
      */
@@ -33,37 +33,33 @@ interface SessionStorageInterface
     /**
      * Checks if the session is started.
      *
-     * @return bool True if started, false otherwise
+     * @return bool
      */
     public function isStarted();
 
     /**
      * Returns the session ID.
      *
-     * @return string The session ID or empty
+     * @return string
      */
     public function getId();
 
     /**
      * Sets the session ID.
-     *
-     * @param string $id
      */
-    public function setId($id);
+    public function setId(string $id);
 
     /**
      * Returns the session name.
      *
-     * @return mixed The session name
+     * @return string
      */
     public function getName();
 
     /**
      * Sets the session name.
-     *
-     * @param string $name
      */
-    public function setName($name);
+    public function setName(string $name);
 
     /**
      * Regenerates id that represents this storage.
@@ -77,24 +73,24 @@ interface SessionStorageInterface
      * only delete the session data from persistent storage.
      *
      * Care: When regenerating the session ID no locking is involved in PHP's
-     * session design. See https://bugs.php.net/bug.php?id=61470 for a discussion.
+     * session design. See https://bugs.php.net/61470 for a discussion.
      * So you must make sure the regenerated session is saved BEFORE sending the
      * headers with the new ID. Symfony's HttpKernel offers a listener for this.
      * See Symfony\Component\HttpKernel\EventListener\SaveSessionListener.
      * Otherwise session data could get lost again for concurrent requests with the
      * new ID. One result could be that you get logged out after just logging in.
      *
-     * @param bool $destroy  Destroy session when regenerating?
-     * @param int  $lifetime Sets the cookie lifetime for the session cookie. A null value
-     *                       will leave the system settings unchanged, 0 sets the cookie
-     *                       to expire with browser session. Time is in seconds, and is
-     *                       not a Unix timestamp.
+     * @param bool     $destroy  Destroy session when regenerating?
+     * @param int|null $lifetime Sets the cookie lifetime for the session cookie. A null value
+     *                           will leave the system settings unchanged, 0 sets the cookie
+     *                           to expire with browser session. Time is in seconds, and is
+     *                           not a Unix timestamp.
      *
-     * @return bool True if session regenerated, false if error
+     * @return bool
      *
      * @throws \RuntimeException If an error occurs while regenerating this storage
      */
-    public function regenerate($destroy = false, $lifetime = null);
+    public function regenerate(bool $destroy = false, ?int $lifetime = null);
 
     /**
      * Force the session to be saved and closed.
@@ -117,13 +113,11 @@ interface SessionStorageInterface
     /**
      * Gets a SessionBagInterface by name.
      *
-     * @param string $name
-     *
      * @return SessionBagInterface
      *
      * @throws \InvalidArgumentException If the bag does not exist
      */
-    public function getBag($name);
+    public function getBag(string $name);
 
     /**
      * Registers a SessionBagInterface for use.
